sysutil: non-Linux OS support.
Use the same implementation as for Linux on non-Linux Unix-like
OSes, such as NetBSD and Illumos. IsLinux will return true for these,
but that's probably fine -- close enough :)

For the clipboard package, the tests fail on the console because
xclip and xsel need an X session to run. Add a check for that and
mark the clipboard as unavailable on Unix if DISPLAY is unset.